+++ RELEASENOTES.md
@@ -4,6 +4,9 @@
 ## Release 2.28.0 - Upcoming
 
 - Updated `@salesforce-ux/design-system-primitive-tokens` to version `0.3.7` to apply BLUE_40 as the link text color.
+  
+- Styling Hooks
+  - Include a new scope for global and shared hooks targeting Agentic Setup elements. This is a continuation from the changes included in 2.27.0.
 
 ## Component Blueprints
 ### [Checkbox Toggle](https://www.lightningdesignsystem.com/components/checkbox-toggle)
@@ -13,6 +16,7 @@
 ### [Combobox](https://www.lightningdesignsystem.com/components/combobox)
 #### Fixed
 - Adds css to fix the height of empty list option in cases where `slds-truncate` class is not present
+#### Fixed
 - Fixed a11y concern with truncated menu options for longer text. The text for each option now wraps by default.
 - Update examples to remove `slds-truncate` class from options. The `slds-truncate` class can still be used to achieve the previous behavior.